pyproject.toml
[project]
name = "malaysia-permits-map"
version = "0.1.0"
description = "MBJB-first municipal development map for Malaysia."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "beautifulsoup4>=4.13.0",
  "duckdb>=1.4.0",
  "geopandas>=1.1.0",
  "geoalchemy2>=0.17.0",
  "httpx>=0.28.0",
  "orjson>=3.11.0",
  "polars>=1.31.0",
  "psycopg[binary]>=3.2.0",
  "pyarrow>=21.0.0",
  "pyogrio>=0.11.0",
  "python-dotenv>=1.1.0",
  "shapely>=2.1.0",
  "sqlalchemy>=2.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
